add csrf header to actions
