+++ src/Utils/getCookie.jsx
@@ -1,0 +1,8 @@
+const getCookie = (name) => {
+    return document.cookie
+        .split('; ')
+        .find((row) => row.startsWith(name + '='))
+        ?.split('=')[1];
+};
+
+export default getCookie;
